feat: Add Dockerfiles for containerizationc
This commit introduces Dockerfiles to containerize the website, backend,
and frontend components of the project.

The Dockerfile for each component defines the necessary environment,
dependencies, and build steps to create a Docker image. This enables
consistent and reproducible deployments across different environments.

- Website: Uses a multi-stage build with Node.js for building and Nginx
  for serving static content.
- Backend: Uses a multi-stage build with the Go toolchain for building
  the application and a minimal Alpine base for runtime.
- Frontend: Uses a multi-stage build with Node.js for building and Nginx
  for serving static content.

These Dockerfiles lay the groundwork for easier deployment and
management
of the project using Docker.